refactor!: move exceptions to proper namespaces
